Fixes uploading of the csv file
missing part, was that the filehandle was already claimed by the write
mehtod and I tried to open it a second time as read-only from another
mehtod. This resulted in empty data, as the content was not yet flushed
to disk.

httpbin.org saved my life as it responded an empty data field in my HTTP
request.